+++ server/models/expense_category.go
@@ -1,8 +1,8 @@
 package models
 
 type ExpenseCategory struct {
-	Id     int     `json:"id" db:"id"`
-	Name   string  `json:"name" db:"name" binding:"required"`
-	Total  float64 `json:"total" db:"total" binding:"required"`
-	UserId int     `json:"user_id" db:"user_id" binding:"required"`
+	Id     int    `json:"id" db:"id"`
+	Name   string `json:"name" db:"name" binding:"required"`
+	Total  int64  `json:"total" db:"total" binding:"required"`
+	UserId int    `json:"user_id" db:"user_id" binding:"required"`
 }
