+++ server/services/balance.go
@@ -2,7 +2,7 @@ package services
 
 import "finance-tracker-server/repository"
 
-func GetBalance(eor *repository.ExpenseOverviewRepository, dr *repository.DepositRepository, userId int) (float64, error) {
+func GetBalance(eor *repository.ExpenseOverviewRepository, dr *repository.DepositRepository, userId int) (int64, error) {
 	expenses, err := eor.GetExpenseSumByUser(userId)
 	if err != nil {
 		return 0, err
